refactor: implementing user id verification and user Role verification in User use cases.

+++ src/main/java/com/TuWebYa/horno/user/application/usecase/DeleteUserUseCaseImpl.java
@@ -1,6 +1,7 @@
 package com.TuWebYa.horno.user.application.usecase;
 
 import com.TuWebYa.horno.user.application.command.DeleteUserCommand;
+import com.TuWebYa.horno.user.application.exception.UserForbiddenException;
 import com.TuWebYa.horno.user.application.port.in.DeleteUserUseCase;
 import com.TuWebYa.horno.user.application.port.out.UserRepositoryPort;
 import org.springframework.stereotype.Service;
@@ -16,6 +17,11 @@ public class DeleteUserUseCaseImpl implements DeleteUserUseCase {
 
     @Override
     public Mono<Void> deleteUser(DeleteUserCommand command) {
+        if (!command.authenticatedUserId().equals(command.id())
+                && command.authenticatedUserRole().equals("USER")) {
+            return Mono.error(new UserForbiddenException());
+        }
+
         return userRepositoryPort.deleteById(command.id());
     }
 }
